Move test_interval/test_dtypes.py to new directory structure
