+++ python/cudf/cudf/tests/series/test_attributes.py
@@ -210,6 +210,13 @@ def test_series_hasnans(data):
     assert gs.hasnans == ps.hasnans
 
 
+def test_category_dtype_attribute():
+    psr = pd.Series(["a", "b", "a", "c"], dtype="category")
+    sr = cudf.Series(["a", "b", "a", "c"], dtype="category")
+    assert isinstance(sr.dtype, cudf.CategoricalDtype)
+    assert_eq(sr.dtype.categories, psr.dtype.categories)
+
+
 def test_dtype_dtypes_equal():
     ser = cudf.Series([0])
     assert ser.dtype is ser.dtypes
